Two changes
Changed the title and description of the sample todos that load when the app is first started. Also added two default projects that load in initially along with the todos.

Changed the way new projects are rendered. Anytime a user adds a new project, the content-container will reflect this change and swap over to the new project.